split uber-script to small ones
